% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics.R
\name{calc_metrics}
\alias{calc_metrics}
\title{Easy function for calculating all polar meatrics.}
\usage{
calc_metrics(input, t = NULL, yr_type, spc, lcut, hcut, return.vecs)
}
\arguments{
\item{input}{A vector of values or a column of an xts object indexed with time stamps.}

\item{t}{An optional vector of time values (e.g., days) corresponding to the input vector sampling points. If \code{input} is an xts object then omit t argument.}

\item{yr_type}{Argument specifying either 'cal_yr' for output (of timing variables) given in days starting from Jan. 1, or 'offset_yr' for output in days starting from the average seasonal minimum.}

\item{spc}{Integer value specifying the number of samples per cycle (measurements per year) in input.}

\item{lcut}{Numeric value in the range [0,0.5] passed to \code{window_idx} function. Indicates the percentile to truncate from the first half of each cycle. For example, 0.15 will result in remove the interval corresponding to [0\%,15\%] representing a window that begins after the 15th-percentile is crossed.}

\item{hcut}{Numeric value in the range (0.5,1] passed to \code{window_idx} function. Indicates the percentile to truncate from the last half of each cycle. For example, 0.85 will result in remove the interval corresponding to [85\%,100\%] representing a window that begins after the 85th-percentile is crossed.}

\item{return.vecs}{logical argument specifying whether or not to include all of the horizontal and vertical component vectors in output.}
}
\value{
Returns a list with the derived polar metrics (e.g., length of season). Optional objects in the list are the actual vecturs used in deriving the polar metrics and the average vector angle and magnitude.
}
\description{
A basic, but fast function for calculating all polar metrics from XTS arrays.
}
\details{
\code{calc_metrics} runs through the entire polar transformation process and conveniently outputs the final polar metrics for all years included in the input.
}
\references{
Brooks, B.J., Lee, D.C., Desai, A.R., Pomara, L.Y.,
  Hargrove, W.W. (in composition). Quantifying seasonal patterns in
  disparate environmental variables using the PolarMetrics R package.
}
\author{
Danny C. Lee, William W. Hargrove, Bjorn J. Brooks, Lars Y. Pomara
}
