% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2d.R
\name{rad2d}
\alias{rad2d}
\title{Transform from radians to time units}
\usage{
rad2d(r, dpc)
}
\arguments{
\item{r}{A vector of numeric values of angles in radians (representing
times/dates) to be converted back to time units.}

\item{dpc}{A numeric value representing the number of divisions per
cycle (e.g., hours per day, days per year, etc.)}
}
\value{
Returns a vector of cyclic dates in the interval [0,dpc+1)
  having the same length as input vector \eqn{r}.
}
\description{
Transform radians to dates in time units (e.g., day of year).
}
\details{
\code{rad2d} transforms input values of \code{r} from radians
  to dates in the interval [0,dpc+1).
}
\examples{
dpy <- 365                 # Days/year
data(mndvi)                # Load data
t <- as.vector(mndvi$day)  # Days since January 1, 2000
r <- t2rad(t, dpc=dpy)     # Transform days of year to radians
days_from_rads <- rad2d(r, dpc=dpy) # Transform radians back to days of year
head(cbind(t, days_from_rads)) # Compare results. Cols should be equivalent
}
\author{
Bjorn J. Brooks, Danny C. Lee, William W. Hargrove, Lars Y. Pomara
}
\references{
Brooks, B.J., Lee, D.C., Desai, A.R., Pomara, L.Y.,
  Hargrove, W.W. (accepted). Quantifying seasonal patterns in
  disparate environmental variables using the PolarMetrics R package.
}

