% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2idx.R
\name{rad2idx}
\alias{rad2idx}
\title{Transform radians to sequential index values}
\usage{
rad2idx(r, spc)
}
\arguments{
\item{r}{A vector of numeric values representing dates in radians.}

\item{spc}{A numeric value indicating the number of samples per cycle
(e.g., samples per year).}
}
\value{
Returns the index values corresponding to \eqn{r}.
}
\description{
Transform radian values to the nearest corresponding
  indices of the original data.
}
\details{
rad2idx finds the corresponding indices for each element of \eqn{r}.
}
\examples{
dpy <- 365                 # Days/year
spy <- 46                  # Samples/year (or samples/cycle)
data(mndvi)                # Load data
t <- as.vector(mndvi$day)  # Days since January 1, 2000
r <- t2rad(t, dpc=dpy)     # Transform days of year to radians
r_idx <- rad2idx(r, spc=spy) # Indices corresponding to radial angles in r
head(cbind(t,r,r_idx))     # Compare results
}
\references{
Brooks, B.J., Lee, D.C., Desai, A.R., Pomara, L.Y.,
  Hargrove, W.W. (2017). Quantifying seasonal patterns in
  disparate environmental variables using the PolarMetrics R package.
}
\author{
Bjorn J. Brooks, Danny C. Lee, William W. Hargrove, Lars Y. Pomara
}
