% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecx.R
\name{vec.x}
\alias{vec.x}
\title{Calculate horizontal component vectors}
\usage{
vec.x(r, v)
}
\arguments{
\item{r}{A vector of numeric values representing dates in radians.}

\item{v}{A vector of numeric values representing a uniformly sampled
series of data that vary over multiple cycles.}
}
\value{
Returns the length of the horizontal vector component for
  each polar coordinate pair.
}
\description{
Calculate the horizontal vector component for each pair
  of polar coordinates.
}
\details{
vec.x calculates the length of the horizontal vector component
  of each polar coordinate pair. Given an angle \eqn{r} in radians and
  a corresponding amplitude \eqn{v} the horizontal vector component,
  \eqn{VX}, is calculated by:
\deqn{VX(r, v) = v * \cos(r)}{VX(r, v) = v * cos(r)}.
}
\examples{
dpy <- 365                 # Days/yr
data(mndvi)                # Load data
t <- as.vector(mndvi$day)  # Days since January 1, 2000
r <- t2rad(t,dpy)          # Transform days of year to radians
v <- as.vector(mndvi$wc)   # MODIS NDVI for Willow Creek tower, WI
vx <- vec.x(r,v)           # The horizontal vector components
}
\references{
Brooks, B.J., Lee, D.C., Desai, A.R., Pomara, L.Y.,
  Hargrove, W.W. (2017). Quantifying seasonal patterns in
  disparate environmental variables using the PolarMetrics R package.
}
\author{
Bjorn J. Brooks, Danny C. Lee, William W. Hargrove, Lars Y. Pomara
}
