% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecy.R
\name{vec.y}
\alias{vec.y}
\title{Calculate vertical component vectors}
\usage{
vec.y(r, v)
}
\arguments{
\item{r}{A vector of numeric values representing dates in radians.}

\item{v}{A vector of numeric values representing a uniformly sampled series of data that vary over multiple cycles.}
}
\value{
Returns the length of the vertical vector component for each polar coordinate pair.
}
\description{
Calculate the vertical vector component for each pair of polar coordinates.
}
\details{
vec.y calculates the length of the vertical vector component of each polar coordinate pair. Given an angle \eqn{r} in radians and a corresponding amplitude \eqn{v} the vertical vector component, \eqn{Vy}, is calculated by:
\deqn{VY(r, v) = v * \sin(r)}{VY(r, v) = v * sin(r)}.
}
\examples{
dpy <- 365                 # Days/yr
data(mndvi)                # Load data
t <- as.vector(mndvi$day)  # Days since January 1, 2000
r <- t2rad(t,dpy)          # Transform days of year to radians
v <- as.vector(mndvi$wc)   # MODIS NDVI for Willow Creek tower, WI
vy <- vec.y(r,v)           # The vertical vector components
}
\references{
Brooks, B.J., Lee, D.C., Desai, A.R., Pomara, L.Y.,
  Hargrove, W.W. (in composition). Quantifying seasonal patterns in
  disparate environmental variables using the PolarMetrics R package.
}
\author{
Danny C. Lee, William W. Hargrove, Bjorn J. Brooks, Lars Y. Pomara
}
